Move content-scope to submodule
