+++ .gitmodules
@@ -4,3 +4,6 @@
 [submodule "shared/tracker-surrogates"]
 	path = shared/tracker-surrogates
 	url = https://github.com/duckduckgo/tracker-surrogates.git
+[submodule "shared/content-scope-scripts"]
+	path = shared/content-scope-scripts
+	url = https://github.com/jonathanKingston/content-scope-scripts
