+++ package.json
@@ -8,9 +8,8 @@
     "repository": "duckduckgo/duckduckgo-privacy-extension",
     "scripts": {
         "install-ci": "npm ci --ignore-scripts && npm rebuild puppeteer && npm run postinstall",
-        "postinstall": "git submodule update --init --recursive && npm run copy-sjcl && export PATH='$PATH:$PWD/node_modules/chromedriver/bin'",
+        "postinstall": "git submodule update --init --recursive && cd shared/content-scope-scripts && npm install && export PATH='$PATH:$PWD/node_modules/chromedriver/bin'",
         "bundle-config": "node scripts/bundleConfig.mjs",
-        "copy-sjcl": "cd node_modules/sjcl/ && ./configure --compress=none --without-all --with-hmac --with-codecHex && make && cp sjcl.js ../../shared/js/content-scope/",
         "build": "echo 'Try npm run release-firefox or npm run release-chrome instead' && exit 0",
         "dev": "echo 'Try npm run dev-firefox or npm run dev-chrome instead' && exit 0",
         "release": "grunt build",
@@ -36,9 +35,6 @@
         "@babel/polyfill": "^7.10.4",
         "@babel/preset-env": "^7.16.7",
         "@fingerprintjs/fingerprintjs": "^3.3.2",
-        "@rollup/plugin-commonjs": "^21.0.1",
-        "@rollup/plugin-node-resolve": "^13.1.2",
-        "@rollup/plugin-replace": "^3.0.1",
         "asana": "^0.18.6",
         "babelify": "10.0.0",
         "brfs": "^2.0.2",
@@ -67,8 +63,6 @@
         "node-fetch": "^3.1.0",
         "puppeteer": "13.0.1",
         "request": "^2.88.2",
-        "rollup": "^2.62.0",
-        "rollup-plugin-dynamic-import-variables": "^1.1.0",
         "sass": "^1.45.2",
         "standard": "16.0.4",
         "timekeeper": "^2.1.1",
@@ -91,7 +85,6 @@
         "punycode": "2.1.1",
         "runtimer": "1.0.3",
         "seedrandom": "^3.0.5",
-        "sjcl": "^1.0.8",
         "tldts": "^5.7.58",
         "underscore": "1.13.2",
         "web-ext": "^6.6.0"
