--- shared/js/content-scope/fingerprinting-hardware-protection.js
@@ -1,19 +1,0 @@
-import { overrideProperty } from './utils'
-
-export function init (args) {
-    overrideProperty('keyboard', {
-        object: Navigator.prototype,
-        origValue: navigator.keyboard,
-        targetValue: undefined
-    })
-    overrideProperty('hardwareConcurrency', {
-        object: Navigator.prototype,
-        origValue: navigator.hardwareConcurrency,
-        targetValue: 2
-    })
-    overrideProperty('deviceMemory', {
-        object: Navigator.prototype,
-        origValue: navigator.deviceMemory,
-        targetValue: 8
-    })
-}
