--- shared/js/content-scope/floc-protection.js
@@ -1,9 +1,0 @@
-export function init () {
-    if ('interestCohort' in Document.prototype) {
-        try {
-            delete Document.prototype.interestCohort
-        } catch {
-            // Throw away this exception, it's likely a confict with another extension
-        }
-    }
-}
